+++ tests/testthat/test-candlestick.R
@@ -5,3 +5,7 @@ ifelse(interactive(), source("tests/cX-function.R"), source("../cX-function.R"))
 test_that("cXcandlestick1", {
     warning("candlestick - plot function missing")
 })
+
+test_that("cXcandlestick2", {
+    warning("candlestick - plot function missing")
+})