+++ tests/testthat/test-heatmap.R
@@ -90,3 +90,12 @@ test_that("cXheatmap11", {
     expect_s3_class(result, "htmlwidget")
 })
 
+test_that("cXheatmap12", {
+    result <- cXheatmap12()
+    if (interactive()) { print(result) }
+    
+    message('middle section too small at default height - working as designed')
+    expect_s3_class(result, "canvasXpress")
+    expect_s3_class(result, "htmlwidget")
+})
+